♻️ refactor: add "mergeClassNames" function and adjust prettier on lint
