+++ .eslintrc.json
@@ -15,9 +15,16 @@
         "plugin:prettier/recommended"
 ],
   "rules": {
+      "react/react-in-jsx-scope": "off",
       "react-hooks/rules-of-hooks": "error",
       "react-hooks/exhaustive-deps": "warn",
-      "react/prop-types": "off"
+      "react/prop-types": "off",
+      "prettier/prettier": [
+        "error",
+        {
+            "endOfLine": "auto"
+        }
+    ]
   },
   "settings": {
       "react": {
