+++ .prettierrc.json
@@ -3,5 +3,6 @@
   "trailingComma": "all",
   "singleQuote": true,
   "printWidth": 120,
-  "tabWidth": 4
+  "tabWidth": 4,
+  "endOfLine": "auto"
 }
