+++ package.json
@@ -5,8 +5,7 @@
     "dependencies": {
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
-        "react-scripts": "5.0.0",
-        "typescript": "^4.5.4"
+        "react-scripts": "5.0.0"
     },
     "scripts": {
         "start": "react-scripts start",
@@ -57,6 +56,7 @@
         "sass-loader": "^12.4.0",
         "style-loader": "^3.3.1",
         "terser-webpack-plugin": "^5.3.0",
+        "typescript": "^4.5.4",
         "ts-loader": "^9.2.6",
         "ts-node": "^10.4.0",
         "webpack": "^5.66.0",
