+++ src/index.ts
@@ -1,2 +1,2 @@
-export { Button } from './button';
-export { Input } from './input';
+export { Button } from './components/button';
+export { Input } from './components/input';
