+++ src/services/classes.ts
@@ -1,0 +1,5 @@
+const mergeClassNames = (classes: string[]): string => {
+    return classes.filter(Boolean).join(' ');
+};
+
+export { mergeClassNames };
