+++ src/stories/Page.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Header } from './Header';
-import './page.css';
+import './page.scss';
 
 export interface PageProps {
     user?: object;
